fix: organize privacy page to be server-side
